bench/requirements.txt
swebench==1.1.2
datasets>=2.19.0
GitPython>=3.1.43
tqdm>=4.66.0
